MAJOR: peers: Deal with updates per bucket
updates are now handled by bucket. It means there is an update list per
bucket. The update lock was also moved in buckets. And it is the real
purpose of this patch. This should reduce contention on this lock.

At the peer level, on the shared table, <last> and <end> markers are now
also per bucket.