BUG/MINOR: check: only try connection reuse for http-check rulesets
In 3.2, a new server keyword "check-reuse-pool" has been introduced. It
allows to reuse a connection for a new check, instead of always
initializing a new one.

Check reuse was only performed if the related keyword is specified and
also if the check does not rely on connection parameters which differ
from the server.

This patch reduces the check reuse to only http-check rulesets. For
other type of checks, reuse must not be performed. Indeed, reuse relies
on HTTP request/response exchanges to ensure the connection is always in
a known state prior to its reutilization.

The impact of this bug is unknown and depends on the proxy/server
configuration. In the case of an HTTP backend with non-HTTP checks,
check-reuse-pool would probably cause a drop in reuse rate.

This must be backported up to 3.2.