MEDIUM: check: do not use QUIC if any connection setting specified
When performing server checks, by default connection settings from the
server are reused. This is also performed for QUIC servers, which allow
to execute QUIC health checks.

This patch adjusts health checks on QUIC servers when at least one check
connection setting is specified, for example the check port. In this
case, consider that the user does not want to use QUIC and fallback to
the TCP transport layer.

This solution is the easiest way to have a consitent behavior for QUIC
servers with checks. Without it, it is possible to encounter process
crashes. It may evolve in the future to be able to perform QUIC checks
with different connection settings.