BROKEN/REGTESTS: ssl_reuse: add explicit 0-rtt tests
BROKEN: works pretty fine alone, but once launched with the whole set
tests, bugs saying "wrong FD type 25 and 26", indicating a bug in vtest
and possibly a mix of FDs between running tests. Dropping VTC_SOCK_TYPE
goes a bit further but issues 503...

New proxies are configured to relay traffic to existing instances
using various combinations:
  - all servers have "allow-0rtt" enabled
  - those which don't have a retry-on must not use 0-rtt
  - those which target a TLSv1.2 server must not use 0-rtt
  - other ones must use it, both in H1 and H2

The number of loops for the server had to be increased a lot since
we're adding 12 tests of 21 connections each.